refactor : 컴포넌트 파스칼 케이스로 변경
